+++ packages/plugins/rsbuild-plugin-extension-monitor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @addfox/rsbuild-plugin-extension-monitor
 
+## 0.2.7
+
+### Patch Changes
+
+- feat: MV2 support, config-level test field, and leaner dev output
+- Updated dependencies
+  - @addfox/common@0.2.6
+  - @addfox/core@0.2.7
+
 ## 0.2.6
 
 ### Patch Changes